undo connection module
