+++ singer_sdk/__init__.py
@@ -4,8 +4,13 @@ from singer_sdk import streams
 from singer_sdk.mapper_base import InlineMapper
 from singer_sdk.plugin_base import PluginBase
 from singer_sdk.sinks import BatchSink, RecordSink, Sink, SQLSink
-from singer_sdk.sql import SQLConnector
-from singer_sdk.streams import GraphQLStream, RESTStream, SQLStream, Stream
+from singer_sdk.streams import (
+    GraphQLStream,
+    RESTStream,
+    SQLConnector,
+    SQLStream,
+    Stream,
+)
 from singer_sdk.tap_base import SQLTap, Tap
 from singer_sdk.target_base import SQLTarget, Target
