--- singer_sdk/sql/__init__.py
@@ -1,4 +1,0 @@
-"""Module for helpers common to SQL streams/sinks."""
-from .connector import SQLConnector
-
-__all__ = ["SQLConnector"]
