+++ singer_sdk/streams/__init__.py
@@ -1,10 +1,9 @@
 """SDK for building Singer taps."""
 
-from singer_sdk.sql import SQLConnector
 from singer_sdk.streams.core import Stream
 from singer_sdk.streams.graphql import GraphQLStream
 from singer_sdk.streams.rest import RESTStream
-from singer_sdk.streams.sql import SQLStream
+from singer_sdk.streams.sql import SQLConnector, SQLStream
 
 __all__ = [
     "Stream",
